add ddhq embedding code
